+++ src/Symfony/Bundle/DebugBundle/DependencyInjection/DebugExtension.php
@@ -11,11 +11,13 @@
 
 namespace Symfony\Bundle\DebugBundle\DependencyInjection;
 
+use Symfony\Bundle\DebugBundle\Command\ServerDumpPlaceholderCommand;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\Extension;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
 use Symfony\Component\DependencyInjection\Reference;
+use Symfony\Component\VarDumper\Dumper\ServerDumper;
 
 /**
  * DebugExtension.
@@ -43,11 +45,18 @@ class DebugExtension extends Extension
         if (null === $config['dump_destination']) {
             //no-op
         } elseif (0 === strpos($config['dump_destination'], 'tcp://')) {
+            $serverDumperHost = $config['dump_destination'];
             $container->getDefinition('debug.dump_listener')
                 ->replaceArgument(1, new Reference('var_dumper.server_dumper'))
             ;
-            $container->getDefinition('var_dumper.command.server_dump')
-                ->replaceArgument(1, new Reference('var_dumper.server_dumper'))
+            $container->getDefinition('data_collector.dump')
+                ->replaceArgument(4, new Reference('var_dumper.server_dumper'))
+            ;
+            $container->getDefinition('var_dumper.dump_server')
+                ->replaceArgument(0, $serverDumperHost)
+            ;
+            $container->getDefinition('var_dumper.server_dumper')
+                ->replaceArgument(0, $serverDumperHost)
             ;
         } else {
             $container->getDefinition('var_dumper.cli_dumper')
@@ -56,6 +65,13 @@ class DebugExtension extends Extension
             $container->getDefinition('data_collector.dump')
                 ->replaceArgument(4, new Reference('var_dumper.cli_dumper'))
             ;
+        }
+
+        if (!isset($serverDumperHost)) {
+            $container->getDefinition('var_dumper.command.server_dump')->setClass(ServerDumpPlaceholderCommand::class);
+            if (!class_exists(ServerDumper::class)) {
+                $container->removeDefinition('var_dumper.command.server_dump');
+            }
         }
     }